Rearrange command function files to prepare for abstraction refactor
