+++ src/commands/clean/cleanFunctions.ts
