+++ src/commands/transcribe/transcribeFunctions.ts
